Rename node class for `JobCalculation` (#2201)
* Rename node class for `JobCalculation`

The node class changed from `JobCalculation` to `CalcJobNode`.

Since the `JobCalculation` class is heavily used by users, as it is the
class that is sub classed to define a job calculation, we cannot quite get
rid of it. Instead we moved all the logic from the old `Calculation` and
`JobCalculation` classes to their new equivalents `ProcessNode` and `CalcJobNode`,
and put an alias to `CalcJobNode` for `JobCalculation`.

This last class can now still be imported from `aiida.orm.calculation` and
`aiida.orm.calculation.job`. Note that the paths that import directly from the
`implementation.general` sub module have been removed, because the job calculation
classes are now fully concrete.

Furthermore, the old entry points for the base calculation classes `inline`, `job`
`work` and `function` have been removed as they have been replaced with the entry
points in the `aiida.node` group for the new node classes.

* Minor changes to address PR comments