Implement the new `calcfunction` decorator. (#2203)
This `calcfunction` decorator replaces the `make_inline` and `optional_inline`
decorators, which are deprecated. The `calcfunction` operates identical to the
`workfunction` except the former gets a `CalcFunctionNode` to represent itself
in the provenance graph as opposed to the `WorkFunctionNode` of the latter.

The behavior of the `optional_inline` decorator is already intrinsically
contained within the `calcfunction` and `workfunction` decorators. Since under
the hood a `FunctionProcess` will be generated, and each process has an input
port `store_provenace` that can be toggled to `False` to prevent the provenance
from being stored, each `calcfunction` can be run without storing the provenance
by simply passing `store_provenance=False` when calling the function.