Move functionality from `verdi calculation` to `verdi process` and `verdi calcjob`
The functionality that was in `verdi calculation` mostly applied only to calculation
jobs, and not to any other calculations as generated for example with the calcfunction
decorator. Other sub commands were generic enough to be put in `verdi process`. Here
we move this common functionality such as the commands `show` and `logshow` to the
`verdi process` command, renaming `logshow` to `report`.

The remaining commands that apply to job calculations `inputls`, `inputcat`, `outputls`,
`outputcat`, `gotocomputer` and `cleanworkdir` are moved to `verdi calcjob`. The old
commands in `verdi calculation` are all deprecated and automatically forwarded to the
new commands.