Added caching of entity collections and default user
Now if a new collection is created via .objects it is cached for the
next time in a lazy store.  Similarly the default user for a User
collection is cached after being created for the first time.  This may
lead to issues if the default user can change for a given profile during
runtime, however for now we don't support this.