Renaming and reorganizing some `verdi` commands (#2204)
* Move functionality from `verdi calculation` to `verdi process` and `verdi calcjob`

The functionality that was in `verdi calculation` mostly applied only to calculation
jobs, and not to any other calculations as generated for example with the calcfunction
decorator. Other sub commands were generic enough to be put in `verdi process`. Here
we move this common functionality such as the commands `show` and `logshow` to the
`verdi process` command, renaming `logshow` to `report`.

The remaining commands that apply to job calculations `inputls`, `inputcat`, `outputls`,
`outputcat`, `gotocomputer` and `cleanworkdir` are moved to `verdi calcjob`. The old
commands in `verdi calculation` are all deprecated and automatically forwarded to the
new commands.

* Move functionality from `verdi work` to `verdi process`

The majority of functionality in `verdi work` that was not already deprecated
and moved to `verdi process`, such as `status` and `report` are generally
applicable to all process types. Therefore it makes sense to move everything
there and deprecate the `verdi work` command.

* Add the `verdi plugin` command

This should serve as the central command to list registered plugins
for the various entry point groups, instead of each verdi command having
its own dedicated sub command. The existing sub commands for `verdi data`
`verdi calculation` and `verdi work` have been deprecated and piped through
to the new `verdi plugin` command.