Add uniqueness constraints on all UUID fields in SqlAlchemy models
This adds the uniqueness constraint on the UUID fields of the models:

 * Computer
 * Comment
 * Group
 * Node
 * Workflow

A check is performed before the migration is applied, to see whether the
affected tables actually contain any entries with duplicate UUIDs. In
this case a warning is emitted and the migration is stopped. The
infrastructure to deduplicate, that is present all ready for nodes, has
not yet been generalized to arbitrary tables.