Implement the new link types
With the new ORM hierarchy in place, the corresponding link types can be defined:

	CREATE: CalculationNode -> Data
	RETURN: WorkflowNode -> Data
	INPUT_CALC: Data -> CalculationNode
	INPUT_WORK: Data -> WorkflowNode
	CALL_CALC: WorkflowNode -> CalculationNode
	CALL_WORK: WorkflowNode -> WorkflowNode

These rules are encoded in the `validate_link` function in `aiida.orm.utils.links`.
This function is called from `add_link_from`, which will be renamed to `add_incoming`,
and validate the triple (source, link, target) representing the addition of a link
from a source node to a target node.