Return correct ORM AuthInfo class for `JobCalculation._get_authinfo`
The method returned the backend AuthInfo class, which does not have
the `get_transport` method that would be called for example in the
`_get_transport` method of the `JobCalculation` class. The fix is
to wrap the backend entity in the ORM class by calling

    `AuthInfo.from_backend_entity`

Added two regression tests for this bug.