Index incoming link cache of Node on link triple instead of just label
The link triple is defined as the tuple of the source node, link type
and link label of an incoming link into a target node. This concept is
implemented in a named tuple `LinkTriple`. Since the introduction of
link types, the link label of incoming links no longer necessarily needs
to be unique, as long as the triple with the link type and source node
*is* unique.

However, the old implementation of the incoming link cache relied on label
uniqueness, because it used only the label as the key in the cache, which
used a dictionary as its data structure. Here we change that data structure
of the node's internal incoming link cache to be a set of link triples.