Make the `link_label` required in `Node.add_incoming`
The `link_label` used to be optional when specifying a link, as long
as both nodes were already stored. If one of the nodes was unstored
the link had to be stored in the internal cache which was indexed
on the link label. This was already changed to be indexed on the
link triple instead, solving that issue, but the auto label generation
in `Node._add_dblink_from` in the case of no label being defined,
relied on a uniqueness constraint of the link label on the database.
This constraint had already been removed a long time ago when the
`RETURN` link type was introduced. However, the logic of the auto
label generation was kept, even though it would never work. We remove
that functionality here and make the explicit passing of a link label
required in the `add_incoming` and `_add_dblink_from` methods.