Make `verdi` callable without sub command (#2240)
Because the `verdi` group was not marked as callable without
subcommand, the `--version` option was broken as it would throw
a `Missing command` error. Adding `invoke_without_command=True` to
the `verdi` group declaration and making the `--version` an eager
option will ensure it works without specifying a sub command.

A test has been added to test this functionality.