Add an archiving transport task for job processes
A common use case for job calculations is that one wants to keep certain
files after it has finished, since they are necessary for restarts, but
one does not want to retrieve them from the remote since they can be rather
big. Instead, one wants to copy these files from the working directory, which
is often a scratch space, to a more permanent location.

Here we implement the `archive` transport task, that is optionally called after
the job has finished and before the retrieval task is triggered. If on the node
the `archive_list` attribute is defined, the worker will archive certain files
as specified to another folder on the remote and attach it as a remote folder
data to the calculation node.