Allow existing profile to be overridden in `verdi setup` (#2244)
With the `--force` flag a user can suppress the exception that is raised
when one tries to create an existing profile in non-interactive mode. With
the flag the existing profile will be overridden with the specified values
except for the profile UUID which will be kept the same as it is auto
generate and can also not be specified by the user on the command line.