Refactor the function process decorator (#2246)
The `calcfunction` and `workfunction` decorator are essentially
running the same code, with the only difference being the type of
node they use to represent their execution in the provenance graph.
The code has been refactored to a generic `process_function` decorator
which takes a node class as an argument. The `calcfunction` and
`workfunction` decorators are then simply defined by wrapping the
`process_function` with the correct node class, `CalcFunctionNode` and
`WorkFunctionNode` respectively.

The difference in link rules between the two concrete decorators is
also completely defined by those node classes, so the logic of the
`FunctionProcess` class that is dynamically generated by the process
function decorator can be implemented once. The common functionality
of defining and overriding inputs is tested in `test_process_functions`.
A `process_function` should never actually be run, but for testing
purposes we defined a dummy ORM class.

Functionality that is specific to the two concrete decorators, such as
which type of links can be created, is tested individually.