Put in an explicit check raising when calculations call another process
By definition, calculation type processes cannot call other processes.
This was already indirectly guarded against by the link validation
process. If one attempted to add a `CALL` link from a `CalulationNode`,
the link validation would raise a `ValueError`. However, it is more
instructive for the user to catch this problem before attempting to add
the link. Therefore, when setting up the database record for a process,
if a parent process is available, we ensure that it is not a calculation
type process, or we raise an `InvalidOperation` exception. A test is
added for a `calcfunction` calling another `calcfunction`.