Force SQLAlchemy to not de-duplicate rows (#2281)
For the original SqlAlchemy implementation the following equation did not hold:

    qb.count() == len(qb.all())

This was because SqlAlchemy was de-duplicating the query set in the `all()` call.
The breaking of this equality was inconsistent with the Django implementation and
the result when operating directly on the database. The postgres behavior is now
restored by setting `self._query._has_mapper_entities = False` on the SqlA query
object. This is a temporary fix and will soon be released as a public option in a new
version of SqlAlchemy.