Implement the database migrations after the provenance redesign
The migrations consist of two steps:

 * Inferring the node process type of the node type
 * Migrating the node type, deleting illegal links and renaming links

The process type column was added sometime ago, which is used to store the
entry point string of the process class that led to the creation of the node.
This is mostly applicable for job calculations and workflows. Since the node
type is the same for all calculations and workflows, the process type is
necessary to keep the information of the process class that ran it.

Older databases may contain job calculations and workchains that were run before
the code was updated to include the process type. There the information for what
sub class, of especially the JobCalculation, was contained in the node type.
Since in this migration, the node type will be migrated to be just the base
CalcJobNode, with type string `node.process.calculation.calcjob.CalcJobNode.`,
the information of the calculation sub class would be lost. Therefore, before
migrating the node types, we attempt to infer the entry point string from the
node type string and store that in the process type column. Node type strings
that cannot be mapped on a known entry point string will get a fallback process
type string and are logged to a file.

After this is completd, a normal database migration is applied, where illegal
links are deleted, the remaining renamed according to the new rules and names and
the node types are renamed according to the new node ORM class type strings.