Ensure correct types of `LinkTriples` returned by `Node.get_stored_link_triples`
The `Node.get_stored_link_triples` was using `LinkTriple` instances where the
type of the `link_type` element was not of `LinkType` but of string, as it was
directly using the string value returned by the query builder for the `type`
column of the `DbLink` table. By wrapping this value in `LinkType` the returned
`LinkTriples` once again have the correct type for all its elements.