Removed `get_aiida_class`
Replaced all instances with either `get_backend_entity` or `get_orm_entity`
(as appropriate).

Also changed backend group to be independent of the ORM.  Because the
iterator was getting the ORM class and returning that, which required it
to access up to the ORM level.  Now it just returns backend entities
which are in turn converted to ORM entities by the `orm.Group`.  This also
means you can't call `len(group.nodes)` because this is not allowed for
iterators but the same can be achieved with `len(list(group.nodes))`