Updated documentation about the usage of `CodeInfo` (#2276)
Both `CalcInfo` and `CodeInfo` are disscussed already in the two examples
(QE and integer summation). Note about the support of multiple codes for a
single calculation is added. This feature was not clearly documented earlier.
A leftover QE example in the plugin development guide is removed.

Also improved the integer/float summation examples with links to direct user
to general plugin system documentation and online examples.