Various speedups for verdi cmdline
In particular when completing tests.
Thanks to @ltalirz
This fixes #376 (again...).
Fixes include:
- using a new instance of the reentry manager that prevents
  automatic reentry scans
- removed unuseful with_dbenv decorators from 'verdi devel tests' and
  from the TestModuleParamType of click
- prevent unnecessary imports of plumpy that is a bit slow to import

Also, we fixed the place where AIIDADB_PROFILE is set in the
main `verdi` click command, to set it also when a default
profile is used (this was probably the reason of the unneeded
with_dbenv calls)