Add migrations to properly deal with `ProcessCalculation` and `WorkCalculation`
There has been a period, at the very beginning of the move to processes
where the node classes had the type `calculation.process.ProcessCalculation.`
These could represent either the execution of a workfunction or a workchain.

Therefore, we first convert all `ProcessCalculations` to `WorkCalculations`
and in a second step, we convert all `WorkCalculations` that have an attribute
called `function_name` to `WorkFunctionNode`. The only remaining `WorkCalculations`
then have to be `WorkChainNodes`.