Add checks for unexpected links in provenance redesign migration
The migration for the provenance redesign will delete illegal links such
as `return` and `call` links outgoing from calculation nodes. Since the
migration will simply delete them, before we do so, we detect these illegal
links, and if found will print a warning to the console and write the
offending links to a migration log file.

This way, the user could potentially restore the deleted links.