Add test for availability of sudo in quicksetup
Quicksetup needs to access the Postgres database, but does not
know anything about the system setup. It initially tries to
see if it is the postgres superuser, but if it is not, it tries
to become the postgres superuser via sudo. This is the backstop
position and relies on the user having a modern Linux-like system.
If, however, the sudo command is not available to this user at all
(for example, perhaps on an HPC cluser), an error will be raised and
quicksetup will fail ungracefully. This commits add a test to see
if sudo can be found before attempting to use it. If it is not, no
attempt is made, and the final manual setup message is printed.

- Add simple check of the availability of the sudo command to
  the user.
- Add warning message if sudo cannot be found.
- Update click.echo calls to the AiiDA echo function.
- Update selected messages to 'warning' type messages.