Implement `verdi database migrate`
Up till now, the triggering of a database migration happened through
separate interfaces for the two different backends. For Django, if
the database schema did not match the code's, a warning message would
be printed asking the user to execute a stand alone python script that
was a modified Django `manage.py` script. For SqlAlchemy, the user was
automatically prompted with an Alembic migration.

Here we unify the migration operation by making it go through `verdi`
through the endpoint `verdi database migrate`. The error message upon
an outdated database schema and the to be executed commands are now
identical for both backends.