Fixed 1714 expose_inputs deepcopy default value
Allowing deepcopy of a stored node to support this use case.  See the
bug report https://github.com/aiidateam/aiida_core/issues/1741 for
default but simply put if you expose the inputs of a workchain with a
stored default plumpy will try to deepcopy that default value which was
expecting because of a ban on deepcopying stored nodes.