Use `ProfileParamType` in `verdi` and expose in `ctx.obj` (#2345)
The `verdi` entry point now uses the `ProfileParamType` for its `-p`
option, which enables tab-completion for the profile option.
In addition, we add the loaded profile to the `obj` attribute of the
click context `ctx`. This way, when a `verdi` sub commands needs the
active profile, one can simply use the `@click.pass_context` decorator
to access the profile through `ctx.obj.profile`.