Allow deepcopy of stored data nodes (#2347)
The `expose_inputs` of the `ProcessSpec` is implemented by `plumpy` and it
recursively deep copies the ports of the original spec to the target spec. This
includes deep copying any default values that those ports may have, which may
also be a stored `Data` node. Originally we prohibited the deep copying of stored
`Data` nodes, because it may suggest that an actual copy of the node would also
be created in the database, while that is not the case. The decision was made to
simply raise when trying to deep copy a stored data node. However, this breaks
the `exposed_inputs` functionality as it is perfectly legal to set a stored data node
as a port default. Since we have to allow users to set stored data nodes as a port
default, we have to lift the ban of deep copying them.