Docs: update the example provenance graph images in concepts section
The original .svg files are also included. They were generated in
Inkspace but saved as plain SVG files. After that they were exported
as PNG files. With the `convert` Unix utility, the images were
rescaled to 800 px as follows:

    convert filename.png -resize 800 filename.png

This will rescale the image to 800px wide, maintaining the aspect ratio
and writing it to the same file.