Refactor the management of the configuration file (#2349)
The configuration file is now abstracted by the `Config` class.
Loading the configuration through `aiida.manage.load_config` will
now return a `Config` instance which internally keep the contents
of the configuration file of the AiiDA instace. Altering the config
will go through this object.

All configuration related code, such as settings and methods to setup
the configuration folder for a new AiiDA instance, are organized
in the `aiida.manage.configuration` module.