Fix bug in `CalcJobNode.get_desc`
The function still called `CalcJobNode.get_state` with the argument
`from_attribute`, which has recently been removed. Originally, the
state was stored in a separate database table and had a proxy as an
attribute in the attribute table. The calc state table has been
removed, leaving the attribute as the only source of the state.