Implement `verdi config` (#2354)
The new command `verdi config` replaces the various commands under
`verdi devel` that were used to list, get, set and unset configuration
options, that were used to be called "properties". Since the term
"property" is a reserved keyword, it was decided to rename them to
"options", which is also the term used by "git". The interface of
`verdi config` also mirrors that of `git config`. That is to say
to get the value of an option simply call `git config <option_name>`
and to set it `git config <option_name> <option_value>`. To unset
it, the `--unset` flag can be used. Finally, the getting, setting
or unsetting can be applied to a certain "scope", meaning to be
configuration wide or profile specific.

To make the implementation of this command simple and clean, the
bulk of the work was in refactoring the definition, construction and
operation on the configuration of an AiiDA instance. This is now
represented by the `Config` class, through which these configuration
options can be set or unset as well as retrieved.