Simplify loading of node class from type string
The correct node sub class that should be used to load a node from the
database is determined by the type string stored in the type column. The
only classes that can legally exist in a database, and therefore be
loaded, are defined in `aiida-core`. Therefore, using the entry point
system to map the type string onto an actual ORM class is no longer
necessary. We rename the `aiida.plugins.loader.load_plugin` function to
the more correct `load_node_class`, which given a type string, will
return the corresponding ORM node sub class.

Note that the whole machinery around generating type and query strings
and loading the nodes based on them is still somewhat convoluted and
contains hacks for two reasons:

 1) Data is not yet moved within the `aiida.orm.node` sub module and as
    a result gets the `data.Data.` type string, which will not match the
    `node.Node.` type when sub classing in queries.

 2) CalcJobProcesses are defined by sub classing JobCalculation
    Until the user directly define a Process sub class that uses the
    `CalcJobNode` as its node class, exceptions will have to be made.

If these two issues are addressed, a lot of the code around type strings
can be simplified and cleaned up.