Use different version for `pylint-django` for python 2 and 3
Starting from `pylint-django>=2.0` only `pylint>=2.0` is supported which
has dropped support for python 2. As a result we need to depend on
different version of `pylint-django` based on the version of python.

With the new version of `pylint` a lot of new checks are added that had
to be either corrected or disabled. Given that some of these checks only
are valid for `pylint>=2.0`, they will trigger a warning in the
pre-commit for python 2. For this reason we have to temporarily disable
the `bad-option-value` globally in the `.pylintrc`. When support for
python 2 is dropped, this should be removed.