Replace call to deprecated `set_` methods with `set_option` (#2361)
The `Node._set_internal` method was still calling deprecated `set_`
methods for options of the `CalcJobNode` sub class. These should now
be set through the explicit `set_option` method. The `JobProcess`
was likewise violating this rule, when setting the options of the
underlying `CalcJobNode` it is using as a storage record.