Consider `X` to be an unknown species in `CifData.has_unknown_species`
The `CifData.has_unknown_species` return whether the chemical formula
contains any unknown species, which were defined as those species that
are not listed in `aiida.common.constants.elements`. This was intended
to filter out those cif files that contained species that could not be
parsed into a `StructureData` that could actually be used for a
condensed matter calculation, since the presence of an unknown species
would render that impossible.

However, recently, the specie `X` was added to the internal dictionary
of elements, breaking the method. Here we explicitly ignore `X` in the
`has_unknown_species` method, returning its implementation to its
original intention.