Implement the `CifData.has_unknown_atomic_sites` property
This property will verify that all coordinates of all atomic sites can
be properly converted into float values. If just one coordinate cannot
be successfully converted to a float, the cif is said to have unknown
atomic sites, as it will not be able to be parsed into a usable
`StructureData` object for further calculations.