Ensure that configuration options are considered for logging config (#2375)
The configuration knows various options to change the logging
configuration, however, these were not respected for two reasons:

 * Logging configuration was not lazily evaluated
 * Globally configured options were not agglomerated

The first problem was caused by the fact that the logging configuration
is evaluated upon loading the `aiida` module, at which point the profile
is not necessarily loaded yet, causing the `get_config_option` functions
to return the option defaults. The solution is to have the dictionary
lazily evaluated by using lambdas, which are resolved when
`configure_logging` is called. Finally, we make sure this function is
called each time the profile is set.

The second problem arose from the fact that if a profile is defined, the
`get_config_option` only returned the config value if explicitly set for
that profile and otherwise it would return the option default. This
means that if the option was defined globally for the configuration it
was ignored. This is now corrected where if the current profile does not
explicitly define a value for the option but it is globally defined, the
global value is returned.