Add data migrations for `TrajectoryData` nodes
In the previous commit, the ORM for `TrajectoryData` was changed to
start storing the `symbols` as a list in the attributes of the node
instead of a `numpy` array in the repository, to make it possible to
query for the symbols. This requires a data migration which is
implemented in this commit.

Since it involves the deletion of the data from the repository after it
has been copied to the database, the migration is performed in two
stages. The first will copy the data from the repository to the
attributes on the node and the second will delete the data in the
repository. In this way, if an exception occurs during the migration, no
data will be lost.