Framework for migration tests in SQLAlchemy
This fixes #2391. In particular it adds a framework for testing migrations
in SQLAlchemy/Alembic, and ports one test from Django, also to act as an
example and to verify that the migration testing is properly working.

I've also removed old files that were running some tests but actually were just
running on a set of example migrations and not on those of AiiDA.