Fix bug in the setup of a profile through `verdi setup/quicksetup`
When setting up a profile that already exists, `verdi` is supposed
prompt the user whether they want to reuse the information of the
existing profile or if they want to change it. However, instead of
asking to confirm, the command was prompting, causing the user to always
have to reenter the information.