Remove methods of `Node` to replace or remove links
The `replace_*link_from` methods where only implemented because they
were needed by the old job calculation created. Since that system is now
removed, this functionality is no longer needed. Likewise, the methods
to remove links, either from the cache or the database were not used nor
tested and so they are removed.