Implement `CalcJob` process class  (#2389)
This commit can be summarized in three steps:

 * Reimplementation of a job calculation as `Process` called `CalcJob`
 * Changing job calculation to be purely informational
 * Remove the old job calculation mechanics and business logic

* Reimplementation of a job calculation as `Process` called `CalcJob`

The old way of creating a job calculation, was to subclass the
`JobCalculation` class and override the `_use_methods` class method to
define the input nodes and the `_prepare_for_submission` to setup the
input files for the calculation. The problem was that these methods were
implemented on the `Node` class, thus mixing the responsabilities of
running and introspecting the results of a completed calculation.

Here we define the `CalcJob` class, a subclass of `Process`. This class
replaces the old `JobCalculation` and allows a user to defined the
inputs and outputs through the `ProcessSpec`, just as they would do for
a `WorkChain`. Except, instead of defining an `outline`, one should
implement the `prepare_for_submission`, which fulfills the exact same
function as before, only it is now a public method of the `CalcJob`
process class.

* Changing job calculation to be purely informational

Finally, the role of the job calculation state, stored as an attribute
with the key `state` on the `CalcJobNode` has changed significantly.
The original job calculations had a calculation state that controlled
the logic during its lifetime. This was already superceded a long time
ago by the process wrapper that now fully governs the progression of the
calculation. Despite the calculation state no longer being authoritative
during the calculation's lifetime, it was still present. Here we finally
fully remove and only leave a stripped down version. The remaining state
is stored as an attribute and is a sub state while the `CalcJob` process
is in an active state and serves as a more granual state that can be
queried for. This is useful, because the process status, which also
keeps similar information is human readable and doesn't allow for easy
querying.

* Remove methods of `Node` to replace or remove links

The `replace_*link_from` methods where only implemented because they
were needed by the old job calculation created. Since that system is now
removed, this functionality is no longer needed. Likewise, the methods
to remove links, either from the cache or the database were not used nor
tested and so they are removed.

* Prevent adding links from or to sealed nodes

When a `Node` instance is sealed, it should be impossible to add links
to or from it. To enforce this, the `validate_incoming` and
`validate_outgoing` methods are overridden in the `Sealable` mixin that
will raise `ModificationNotAllowed` if the node is sealed.