Implement functionality to export and import `Log` entries. (#2393)
The native logging of python is extended with a `DbLogHandler` that
will, when it is configured for a specific AiiDA entity (a Node, Group
or any other database entity), store a copy of the record in the `DbLog`
table with a reference to the entity it belongs to. Since the entities
do not necessarily live in the same table, a foreign key could not be
used, so instead the primary key was stored as well. However, when
exporting and importing entities and their log messages, the entities
would get new primary keys, which would orphan the log messages. The
attribute that is persistent across exporting and importing is the UUID.
Therefore the reference attribute for log entries to their entities has
been changed from the `pk` to the `uuid`.