Move `aiida.orm.data` within the `aiida.orm.node` module
This final move makes the hierarchy of all node modules and its sub
classes consistent. The `Data` and `ProcessNode` classes who are both
sub classes of the `Node` base class and are direct siblings live on the
same level in the hierarchy, direct within the `aiida.orm.node` module.

Note that because the type string of all `Data` nodes now will start
with `node.` instead of `data.`, a data migration had to be added.