Generalize the UUID deduplication for more database tables
The original deduplication code was implemented only for the `DbNode`
table but has now been extended to `DbComputer` and `DbGroup`. Note that
this has not been done for `DbComment`, because it does not yet have an
entity loader. If `load_comment` were to be implemented, support for
deduplication can be added to the `deduplicate_uuids` function.