Further simplification of node type string definition and loading (#2401)
With the recent introduction of the `CalcJob` process, which now ensures
that also for job calculations, it is the `CalcJobNode` that gets stored
in the database, which has a type string just like all the other node
types. This made a lot of the old logic `aiida.plugins.loader` to
determine the correct `type` string for a `Node` (sub) class obsolete as
now all node classes (except Data) are per definition internal to `aiida-core` and
have a type string that is based directly on their module path.

The few functions that remain, `get_type_string_from_class` that formats
the correct type string for a given `Node` class and the
`load_node_class` which can reload that class from the string have been
moved to `aiida.orm.utils.node` since it has no longer anything to do
with a "plugin".

Finally, by moving the `WorkflowNode`, `CalculationNode` and
`ProcessNode` classes in the `__init__` files of their respective
modules allows us to add further simplification to `load_node_class` as
well as making all these classes and their subclasses importable from
the `aiida.orm.node` module. Eventually, when `aiida.orm.data` is also
moved there, all classes can even be exposed directly on `aiida.orm` and
that is the deepest a user should ever go for they should need or be
allowed to import.