Implement support for process classes to `QueryBuilder.append`
Any `Process` class will have an associated `ProcessNode` sub class
associated with it that it will use to store its provenance when ran. In
addition with the `process_type`, determined by its specific process sub
class entry point, defines a set of nodes that correspond to that
process class.

For example: the `ArithmeticAddCalculation` will use a `CalcJobNode` to
store its executions and has a process type according to its entry point:

    `aiida.calculations:arithmetic.add`

This commit implements the functionality necessary that allows a user to
append the `ArithmeticAddCalculation` process class to a query builder
instance directly.