Fix the wrong migration for the `DbLog` table
In commit `c0fba2e38557fe1ea535d1dc24076cba99212616`, a migration was
introduced for the `DbLog` table in order to support the export and
import of log entries. However, this migration was flawed and would
lose valuable information.

The old `DbLog` was designed to support adding log messages to various
entities. Because these entities were not restricted to just nodes, and
therefore could live in different database tables, a foreign key could
not be used for the relationship, but instead an `objpk` and `objname`
column were used to allow multiple table references. However, for the
import to know whether an exported log record is already present in the
database and therefore should not be imported again, the object pk could
not be used as these do remain the same between databases. To this end
the migration added a UUID column to the `DbLog` table and dropped the
`objpk` column. However, the data was not migrated before doing so and
so the connection between logs and their entities was lost.

After the faulty commit, we realized that the reason for not using the
foreign key was that logs needed to be able to be associated with both
nodes as well as legacy workflows, but since support for the latter had
already been dropped this was no longer necessary. Given that now only
nodes can have associated logs, we can revert the `objpk` to be a proper
foreign key and the `objname` can be dropped. Since the logs for legacy
workflows (and any other unexpected entities) will be dropped from the
database, they should be exported to a file for archival purposed.

In summary, the migrations perform the following logical actions:

 * Export existing logs for legacy worklows and other entities to file
 * Delete records from legacy workflows and other entities from `dblog`
 * Create a foreign key `dbnode_id`
 * Migrate data from `objpk` to `dbnode_id`
 * Create a new column `uuid` for `DbLog` that is unique
 * Delete the `objpk` and `ojbname` columns from `DbLog`
 * Delete `objpk` and `objname` from the `metadata` of `DbLog` records

Note that the addition of the unique `uuid` column had to be done in
three steps. The problem is that the default value for the UUID is
generated by a python function which cannot therefore be done on a
database level but the values have to be set after the creation of the
column. However, as soon as the column is created the uniqueness
constraint is violated since it will be empty for all the existing
records. For that reason the migration is split in three steps. In the
first, the column is created and set as nullable. The second step will
populate the records using the python function to generate the UUIDs and
finally, the third migration will alter the column to be unique.

Note that for Django this migration could be written in a single file,
but for SqlAlchemy it had to be broken into several migrations. This is
because of Alembic, the framework used to perform SqlAlchemy migrations,
which had trouble with setting the UUIDs after creation of the column in
a single transaction. To circumvent this problem, the migration was split
into three separate migrations and Alembic is now instructed to use a
single transaction per revision, mirroring the migration behavior of
Django.